Finish Architecture Refactor
I again had to manual change many namespace, using and file names as Rider couldn't automate the renaming process, however it has been built and ran with no error.
