+++ src/Chirp.Web/Pages/Public.cshtml.cs
@@ -1,5 +1,7 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
+using Chirp.Infrastructure;
+using Chirp.Infrastructure.Services;
 
 namespace Chirp.Web.Pages;
