+++ src/Chirp.Web/Pages/UserTimeline.cshtml.cs
@@ -1,4 +1,7 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Chirp.Infrastructure;
+using Chirp.Infrastructure.Services;
+
+using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 
 namespace Chirp.Web.Pages;
