+++ test/Chirp.Tests/Chirp.Tests.csproj
@@ -21,7 +21,9 @@
 
 
     <ItemGroup>
-      <ProjectReference Include="..\..\src\Chirp.Web\Chirp.Web.csproj" />
+        <ProjectReference Include="..\..\src\Chirp.Web\Chirp.Web.csproj" />
+        <ProjectReference Include="..\..\src\Chirp.Infrastructure\Chirp.Infrastructure.csproj" />
+        <ProjectReference Include="..\..\src\Chirp.Core\Chirp.Core.csproj" />
     </ItemGroup>
